Revert "Enhancement: Pricing Area"
